Add state columns and polygon type, include buurten
State columns denote columns that are added to each entity that
has states. Gebieden have states where each state has a
start date, end date and a sequence number and registration date.

The sequence number and registration date are required for every
entity that has states and are added automatically.

Polygon type is used for gebieden import.

Buurten is the first entity that has been imported from the
gebieden collection