Added a change to the model, so migrations for events was added
The model for kadastralesecties changed while events had already had been imported. Truncating the table and re-applying the events resulted in an error, so the event data had to be changed to fit the new model. A migration file was added which allows for multiple changes on a version. Events will loop try to find the migration if the version of the model is newer and apply the migration to the event data before applying the event to the database.