Also update QA catalog when no issues are present
This solves the case that when all issues have been solved the open issues should be deleted
Formerly empty issues were skipped which resulted in issues remain open.
This is because deletion is derived from the fact that nothing is imported.

In order to prevent 'overflowing' workflow with useless QA-update jobs
only the functional jobs are processed